Support a base icon url prefix
I use data attribute to set my icons, in the form data-jstree='{"icon":"http://jstree.com/tree.png"}'. Unfortunately the real url is by far much longer, and on a tree of more than 10000 items it causes excessive overhead in the page generation time and while transferring the html to the browser.
This pull request introduces an additional optional parameter to be used during the jstree instantiation:

"core" : {
  "themes" : {
    "iconsurl" : "../long/long/base/url/images/"
  }
}

It is backward compatible: when the parameter iconsurl is unset, jstree simply works as usual.