Update jstree.checkbox.js to allow checkbox/selection separation
It is not always desirable to check a checkbox associated with a node when the node is simply clicked upon. In my case, I wish to use this particular event to show additional details about a node. I have therefore added a check_by_selection option which defaults to true. This is used to determine what triggers a node to be checked/unchecked. If it is set to true, clicking anywhere on the node will toggle it. If set to false, only clicking on the checkbox will trigger this.