Added a return statement to the search plugin, so that the search function also returns the found items.
I did this because even if you set 'show_only_matches' to true, a search with no matches would show the whole tree.
Which is a behavior that seems counter-intuitive. Not knowing if that was intended, I simply added the return statement so that one can check if the search result array length is 0 and then simply hide the tree.